refactor(core): rom-aware base-point name decode + explicit int casts (#1464)
Copilot review round 3:
- GetWorldMapPointName decoded via NameResolver.GetTextById, which reads
  CoreState.ROM (the ambient ROM) — wrong name when the passed rom differs.
  Decode against the PASSED rom via new FETextDecode(rom, encoder).Decode(textid)
  + NameResolver.StripControlCodes (mirrors StatusRMenuCore.GetMenuName). Also
  switch the +28 bounds check to the overflow-safe FitsInRom.
- Add explicit (int) casts on rom.Data[...] index expressions in the test setup
  loops. byte[] indexing already accepts uint/long index expressions (so the
  files always compiled + passed), but the casts end the recurring review noise
  and read clearer.

+++ FEBuilderGBA.Avalonia.Tests/GapSweep/MonsterWMapProbabilityParityTests.cs
@@ -234,8 +234,8 @@ public class MonsterWMapProbabilityParityTests
 
         for (int i = 0; i < MonsterWMapProbabilityCore.StageCount; i++)
         {
-            rom.Data[stageEirikaBase + i] = (byte)(0x10 + i);
-            rom.Data[stageEphraimBase + i] = (byte)(0x40 + i);
+            rom.Data[(int)(stageEirikaBase + i)] = (byte)(0x10 + i);
+            rom.Data[(int)(stageEphraimBase + i)] = (byte)(0x40 + i);
         }
 
         return (rom, stageEirikaBase, probEirikaBase);
